requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=1.2.18
langchain-openai>=1.1.10
langchain-anthropic>=0.3.0
langgraph>=0.2.0
numpy<2.0